Rename src package to pkg
Follow Go standard layout:
https://github.com/golang-standards/project-layout
