+++ pkg/fred/common/common.go
