+++ pkg/fred/common/error/error.go
