+++ pkg/fred/series/date.go
