+++ pkg/fred/series/response/response.go
