+++ pkg/investment_score/calculator/calculator_test.go
@@ -2,9 +2,9 @@ package calculator_test
 
 import (
 	"context"
-	"supermarine1377/yebis/src/fred/series/response"
-	"supermarine1377/yebis/src/investment_score/calculator"
-	"supermarine1377/yebis/src/investment_score/calculator/mock"
+	"supermarine1377/yebis/pkg/fred/series/response"
+	"supermarine1377/yebis/pkg/investment_score/calculator"
+	"supermarine1377/yebis/pkg/investment_score/calculator/mock"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
