+++ pkg/investment_score/calculator/mock/calculator.go
@@ -12,7 +12,7 @@ package mock
 import (
 	context "context"
 	reflect "reflect"
-	response "supermarine1377/yebis/src/fred/series/response"
+	response "supermarine1377/yebis/pkg/fred/series/response"
 	time "time"
 
 	gomock "go.uber.org/mock/gomock"
