+++ pkg/investment_score/mock/investment_score.go
