+++ pkg/investment_score/record/record.go
